chore: encapsulate constant folding logic better
